+++ mstrio/project_objects/incremental_refresh_report/__init__.py
@@ -1,4 +1,2 @@
 # flake8: noqa
-from .advanced_properties import *
-from .template import *
 from .incremental_refresh_report import *